restructured cashier-default.spec.js into separate files (components)
